Update to use LifeCycles, reorganize functions, update docs
